test_requirements.txt
pytest==7.3.1
pytest-mock==3.10.0
moto==4.1.3
coverage==7.2.5
boto3-stubs[essential,organizations]==1.26.90
pytest-cov==4.1.0
freezegun==1.2.2
pyjq